+++ src/model/album.rs
@@ -66,9 +66,8 @@ pub struct FullAlbum {
 /// Full Albums wrapped by Vec object
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/albums/get-several-albums/)
-// TODO: Reduce this wrapper object to `Vec<FullAlbum>`
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct FullAlbums {
+#[derive(Deserialize)]
+pub(in crate) struct FullAlbums {
     pub albums: Vec<FullAlbum>,
 }